Refactor tests
* Config-related tests in their own source file
* More grouping of tests under classes to document what they are testing in common
