+++ tests/test_snakefmt.py
@@ -1,11 +1,9 @@
-import os
 import re
 import tempfile
 from collections import Counter
 from pathlib import Path
 from unittest import mock
 
-import click
 import pytest
 from click.testing import CliRunner
 
@@ -15,7 +13,6 @@ from snakefmt.snakefmt import (
     construct_regex,
     main,
     get_snakefiles_in_dir,
-    read_snakefmt_defaults_from_pyproject_toml,
 )
 
 
@@ -24,14 +21,14 @@ def cli_runner() -> CliRunner:
     return CliRunner()
 
 
-class TestCLI:
+class TestCLIBasic:
     def test_noArgsPassed_printsNothingToDo(self, cli_runner):
         params = []
         actual = cli_runner.invoke(main, params)
         assert actual.exit_code == 0
         assert "Nothing to do" in actual.output
 
-    def test_nonExistantParam_nonZeroExit(self, cli_runner):
+    def test_nonExistentParam_nonZeroExit(self, cli_runner):
         params = ["--fake"]
         actual = cli_runner.invoke(main, params)
         assert actual.exit_code != 0
@@ -52,18 +49,6 @@ class TestCLI:
         assert actual.exit_code != 0
         assert "Cannot mix stdin (-) with other files" in actual.output
 
-    def test_invalidIncludeRegex_nonZeroExit(self, cli_runner):
-        params = ["--include", "?", str(Path().resolve())]
-        actual = cli_runner.invoke(main, params)
-        assert actual.exit_code != 0
-        assert "Invalid regular expression" in str(actual.exception)
-
-    def test_invalidExcludeRegex_nonZeroExit(self, cli_runner):
-        params = ["--exclude", "?", str(Path().resolve())]
-        actual = cli_runner.invoke(main, params)
-        assert actual.exit_code != 0
-        assert "Invalid regular expression" in str(actual.exception)
-
     def test_stdinAsSrc_WritesToStdout(self, cli_runner):
         stdin = f"rule all:\n{TAB}input: 'c'"
         params = ["--verbose", "-"]
@@ -75,6 +60,67 @@ class TestCLI:
         expected_output = f'rule all:\n{TAB}input:\n{TAB*2}"c",\n'
 
         assert actual.output == expected_output
+
+    def test_src_dir_arg_files_modified_inplace(self, cli_runner):
+        with tempfile.TemporaryDirectory() as tmpdir:
+            content = 'include: "a"'
+            abs_tmpdir = Path(tmpdir).resolve()
+            snakedir = abs_tmpdir / "workflows"
+            snakedir.mkdir()
+            snakefile = snakedir / "Snakefile"
+            snakefile.write_text(content)
+            params = [str(tmpdir)]
+
+            cli_runner.invoke(main, params)
+            expected_contents = content + "\n"
+            actual_contents = snakefile.read_text()
+
+            assert actual_contents == expected_contents
+
+    def test_file_arg_write_back_happens(self, cli_runner, tmp_path):
+        content = 'include: "a"'
+        file = tmp_path / "Snakefile"
+        file.write_text(content)
+        params = [str(file)]
+        original_stat = file.stat()
+
+        cli_runner.invoke(main, params)
+        actual_stat = file.stat()
+
+        assert actual_stat != original_stat
+
+        actual_content = file.read_text()
+        expected_content = content + "\n"
+
+        assert actual_content == expected_content
+
+    def test_file_arg_file_requires_no_changes_no_write_back_happens(
+        self, cli_runner, tmp_path
+    ):
+        content = 'include: "a"\n'
+        file = tmp_path / "Snakefile"
+        file.write_text(content)
+        params = [str(file)]
+        expected_stat = file.stat()
+
+        cli_runner.invoke(main, params)
+        actual_stat = file.stat()
+
+        assert actual_stat == expected_stat
+
+
+class TestCLIRegex:
+    def test_invalidIncludeRegex_nonZeroExit(self, cli_runner):
+        params = ["--include", "?", str(Path().resolve())]
+        actual = cli_runner.invoke(main, params)
+        assert actual.exit_code != 0
+        assert "Invalid regular expression" in str(actual.exception)
+
+    def test_invalidExcludeRegex_nonZeroExit(self, cli_runner):
+        params = ["--exclude", "?", str(Path().resolve())]
+        actual = cli_runner.invoke(main, params)
+        assert actual.exit_code != 0
+        assert "Invalid regular expression" in str(actual.exception)
 
     def test_config_adherence_for_python_outside_rules(self, cli_runner, tmp_path):
         stdin = "include: 'a'\nlist_of_lots_of_things = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]"
@@ -131,6 +177,8 @@ list_of_lots_of_things = [
 
         assert actual.output == expected_output
 
+
+class TestCLICheck:
     def test_check_file_needs_no_changes_exit_code_0(self, cli_runner):
         stdin = 'include: "a"\n'
         params = ["--check", "-"]
@@ -214,6 +262,23 @@ list_of_lots_of_things = [
         expected_exit_code = CheckExitCode.ERROR
         assert result.exit_code == expected_exit_code
 
+    def test_check_and_diff_only_runs_check(self, cli_runner, tmp_path):
+        content = 'include: "a"\n'
+        file1 = tmp_path / "Snakefile"
+        file1.write_text(content)
+        file2 = tmp_path / "Snakefile2"
+        content += "x='foo'"
+        file2.write_text(content)
+        params = ["--check", "--diff", str(file1), str(file2)]
+
+        result = cli_runner.invoke(main, params)
+
+        expected_exit_code = CheckExitCode.WOULD_CHANGE.value
+        assert result.exit_code == expected_exit_code
+        assert result.output == ""
+
+
+class TestCLIDiff:
     def test_diff_works_as_expected(self, cli_runner):
         stdin = "include: 'a'\n"
         params = ["--diff", "-"]
@@ -290,230 +355,6 @@ list_of_lots_of_things = [
         expected_contents = content
         actual_contents = snakefile.read_text()
         assert actual_contents == expected_contents
-
-    def test_check_and_diff_only_runs_check(self, cli_runner, tmp_path):
-        content = 'include: "a"\n'
-        file1 = tmp_path / "Snakefile"
-        file1.write_text(content)
-        file2 = tmp_path / "Snakefile2"
-        content += "x='foo'"
-        file2.write_text(content)
-        params = ["--check", "--diff", str(file1), str(file2)]
-
-        result = cli_runner.invoke(main, params)
-
-        expected_exit_code = CheckExitCode.WOULD_CHANGE
-        assert result.exit_code == expected_exit_code
-        assert result.output == ""
-
-    def test_file_requires_no_changes_no_write_back_happens(self, cli_runner, tmp_path):
-        content = 'include: "a"\n'
-        file = tmp_path / "Snakefile"
-        file.write_text(content)
-        params = [str(file)]
-        expected_stat = file.stat()
-
-        cli_runner.invoke(main, params)
-        actual_stat = file.stat()
-
-        assert actual_stat == expected_stat
-
-    def test_file_requires_changes_write_back_happens(self, cli_runner, tmp_path):
-        content = 'include: "a"'
-        file = tmp_path / "Snakefile"
-        file.write_text(content)
-        params = [str(file)]
-        original_stat = file.stat()
-
-        cli_runner.invoke(main, params)
-        actual_stat = file.stat()
-
-        assert actual_stat != original_stat
-
-        actual_content = file.read_text()
-        expected_content = content + "\n"
-
-        assert actual_content == expected_content
-
-    def test_src_dir_is_searched_for_files(self, cli_runner):
-        with tempfile.TemporaryDirectory() as tmpdir:
-            content = 'include: "a"'
-            abs_tmpdir = Path(tmpdir).resolve()
-            snakedir = abs_tmpdir / "workflows"
-            snakedir.mkdir()
-            snakefile = snakedir / "Snakefile"
-            snakefile.write_text(content)
-            params = [str(tmpdir)]
-
-            cli_runner.invoke(main, params)
-            expected_contents = content + "\n"
-            actual_contents = snakefile.read_text()
-
-            assert actual_contents == expected_contents
-
-
-class TestReadSnakefmtDefaultsFromPyprojectToml:
-    def test_no_value_passed_and_no_pyproject_changes_nothing(self, tmpdir):
-        os.chdir(tmpdir)
-        default_map = dict()
-        ctx = click.Context(click.Command("snakefmt"), default_map=default_map)
-        param = mock.MagicMock()
-        value = None
-
-        return_val = read_snakefmt_defaults_from_pyproject_toml(ctx, param, value)
-
-        assert return_val is None
-
-        actual_default_map = ctx.default_map
-        expected_default_map = dict()
-
-        assert actual_default_map == expected_default_map
-
-    def test_pyproject_present_but_empty_changes_nothing_returns_pyproject_path(
-        self, tmpdir
-    ):
-        os.chdir(tmpdir)
-        pyproject = Path("pyproject.toml")
-        pyproject.touch()
-        default_map = dict()
-        ctx = click.Context(click.Command("snakefmt"), default_map=default_map)
-        param = mock.MagicMock()
-        value = None
-
-        actual_config_path = read_snakefmt_defaults_from_pyproject_toml(
-            ctx, param, value
-        )
-        expected_config_path = None
-
-        assert actual_config_path == expected_config_path
-
-        actual_default_map = ctx.default_map
-        expected_default_map = dict()
-
-        assert actual_default_map == expected_default_map
-
-    def test_no_value_passed_and_pyproject_present_changes_default_line_length(
-        self, tmpdir
-    ):
-        os.chdir(tmpdir)
-        pyproject = Path("pyproject.toml")
-        pyproject.write_text("[tool.snakefmt]\nline_length = 4")
-        default_map = dict(line_length=88)
-        ctx = click.Context(click.Command("snakefmt"), default_map=default_map)
-        param = mock.MagicMock()
-        value = None
-
-        actual_config_path = read_snakefmt_defaults_from_pyproject_toml(
-            ctx, param, value
-        )
-        expected_config_path = str(pyproject)
-
-        assert actual_config_path == expected_config_path
-
-        actual_default_map = ctx.default_map
-        expected_default_map = dict(line_length=4)
-
-        assert actual_default_map == expected_default_map
-
-    def test_no_value_passed_and_pyproject_present_unknown_param_adds_to_default_map(
-        self, tmpdir
-    ):
-        os.chdir(tmpdir)
-        pyproject = Path("pyproject.toml")
-        pyproject.write_text("[tool.snakefmt]\nfoo = true")
-        default_map = dict()
-        ctx = click.Context(click.Command("snakefmt"), default_map=default_map)
-        param = mock.MagicMock()
-        value = None
-
-        actual_config_path = read_snakefmt_defaults_from_pyproject_toml(
-            ctx, param, value
-        )
-        expected_config_path = str(pyproject)
-
-        assert actual_config_path == expected_config_path
-
-        actual_default_map = ctx.default_map
-        expected_default_map = dict(foo=True)
-
-        assert actual_default_map == expected_default_map
-
-    def test_value_passed_reads_from_path(self, tmpdir):
-        os.chdir(tmpdir)
-        pyproject = Path("snakefmt.toml")
-        pyproject.write_text("[tool.snakefmt]\nfoo = true")
-        default_map = dict()
-        ctx = click.Context(click.Command("snakefmt"), default_map=default_map)
-        param = mock.MagicMock()
-        value = str(pyproject)
-
-        actual_config_path = read_snakefmt_defaults_from_pyproject_toml(
-            ctx, param, value
-        )
-        expected_config_path = str(pyproject)
-
-        assert actual_config_path == expected_config_path
-
-        actual_default_map = ctx.default_map
-        expected_default_map = dict(foo=True)
-
-        assert actual_default_map == expected_default_map
-
-    def test_value_passed_but_default_map_is_empty_still_updates_defaults(self, tmpdir):
-        os.chdir(tmpdir)
-        pyproject = Path("snakefmt.toml")
-        pyproject.write_text("[tool.snakefmt]\nfoo = true")
-        default_map = None
-        ctx = click.Context(click.Command("snakefmt"), default_map=default_map)
-        param = mock.MagicMock()
-        value = str(pyproject)
-
-        actual_config_path = read_snakefmt_defaults_from_pyproject_toml(
-            ctx, param, value
-        )
-        expected_config_path = str(pyproject)
-
-        assert actual_config_path == expected_config_path
-
-        actual_default_map = ctx.default_map
-        expected_default_map = dict(foo=True)
-
-        assert actual_default_map == expected_default_map
-
-    def test_value_passed_in_overrides_pyproject(self, tmpdir):
-        os.chdir(tmpdir)
-        snakefmt_config = Path("snakefmt.toml")
-        snakefmt_config.write_text("[tool.snakefmt]\nfoo = true")
-        pyproject = Path("pyproject.toml")
-        pyproject.write_text("[tool.snakefmt]\n\nfoo = false\nline_length = 90")
-        default_map = dict()
-        ctx = click.Context(click.Command("snakefmt"), default_map=default_map)
-        param = mock.MagicMock()
-        value = str(snakefmt_config)
-
-        actual_config_path = read_snakefmt_defaults_from_pyproject_toml(
-            ctx, param, value
-        )
-        expected_config_path = str(snakefmt_config)
-
-        assert actual_config_path == expected_config_path
-
-        actual_default_map = ctx.default_map
-        expected_default_map = dict(foo=True)
-
-        assert actual_default_map == expected_default_map
-
-    def test_malformatted_toml_raises_error(self, tmpdir):
-        os.chdir(tmpdir)
-        pyproject = Path("pyproject.toml")
-        pyproject.write_text("foo:bar,baz\n{dict}&&&&")
-        default_map = dict()
-        ctx = click.Context(click.Command("snakefmt"), default_map=default_map)
-        param = mock.MagicMock()
-        value = None
-
-        with pytest.raises(click.FileError):
-            read_snakefmt_defaults_from_pyproject_toml(ctx, param, value)
 
 
 class TestConstructRegex:
